refactor: mock data 분리
