--- src/utils/dataCarousel.ts
@@ -1,99 +1,0 @@
-export const before1 = {
-  id: 1,
-  lifeQuote: "“책은 그것을 적절히 선택할 수 있는 독자에게 갖가지의 즐거움을 안겨준다.”",
-  public: "- 몽키스테외 -",
-  header: "방향성을 찾아보세요.",
-  desc: [
-    "어떤 부분에서의 성장을 기대하시나요?",
-    "이 책에 갖는 기대를 설정함으로써 자신만의",
-    "방향성을 가지고 책을 읽어 나갈 수 있을 거예요.",
-  ],
-};
-
-export const before2 = [
-  {
-    id: 1,
-    lifeQuote: "“생각하지 않고 읽는 것은 잘 씹지 않고 먹는 것과 같다.”",
-    public: "- 에드먼드 버크 -",
-    header: "전략을 세워 보세요.",
-    desc: [
-      "책장을 넘기기 전,",
-      "자신의 기대를 채우기 위한 전략을 세워 보아요.",
-      "이때는 책의 제목, 서론, 목차를 보는 게 좋아요.",
-    ],
-  },
-  {
-    id: 2,
-    lifeQuote: "책의 제목, 서론에는 저자가 전달하고자 하는 핵심 메시지가 들어있어요.",
-    public: "",
-    header: "방향성을 찾아보세요.",
-    desc: [
-      "제목, 서론에 드러나는 저자의 핵심 메시지와",
-      "내가 이 책에 기대한 바를 서로 비교해 보세요.",
-      "책과 나의 방향성이 다르면 기대감을 조정하거나,",
-      "새로운 책을 선택할 수 있어요.",
-    ],
-  },
-  {
-    id: 3,
-    lifeQuote: "책의 목차는 핵심 메시지를 효과적으로 전달하기 위해 구성 되었어요.",
-    public: "",
-    header: "기대를 채워줄 부분을 파악해 보세요.",
-    desc: [
-      "저자가 목차를 어떤 순서로 구성했고,",
-      "그렇게 구성한 이유는 무엇인지 파악해 보세요.",
-      "핵심 메시지를 전달하는 흐름 속에서 내 기대를",
-      "채워줄 부분들을 가려낼 수 있을 거예요.",
-    ],
-  },
-];
-
-export const before3 = {
-  id: 1,
-  lifeQuote: "“내가 세계를 알게 된 것은 책에 의해서였다.”",
-  public: "- 사르트르 -",
-  header: "스스로에게 질문을 던져 보세요.",
-  desc: [
-    "제목, 서론, 목차를 살펴보며 호기심이 생긴 것들",
-    "에 대해 스스로에게 질문들을 던져 보세요.",
-    "이 단계에서 만든 질문들은 다음 단계를 거쳐",
-    "더 깊고 넓은 질문과 사색으로 이어질 거예요.",
-  ],
-};
-
-export const after1 = [
-  {
-    id: 1,
-    lifeQuote: "",
-    public: "",
-    header: "본격적으로 책을 읽어봐요!",
-    desc: [
-      "독서 전 단계 잘 마무리하고 오셨나요?",
-      "지금부터 독서 전 단계에서 설정한 방향을 따라",
-      "독서를 해 나갈 거예요.",
-    ],
-  },
-  {
-    id: 2,
-    lifeQuote: "",
-    public: "",
-    header: "내가 만든 질문의 답변을 찾아 보세요.",
-    desc: [
-      "앞 단계에서 자신의 기대를 채우기 위한 질문을 만들었어요.",
-      "이렇게 만든 질문들에 답하다 보면",
-      "지식이 구조화되어 머릿속에 저장 될 거예요.",
-    ],
-  },
-  {
-    id: 3,
-    lifeQuote: "",
-    public: "",
-    header: "답변에 대한 꼬리 질문을 만들어 보세요.",
-    desc: [
-      "독서 전에 만들었던 큰 질문들은",
-      "책장을 넘길수록 더 날카롭고 깊어져 갈 거예요.",
-      "호기심을 느낀다면 언제든,",
-      "꼬리 질문들을 만들어 보세요.",
-    ],
-  },
-];
